requirements.txt
opencv-python==4.8.1.78
face_recognition==1.3.0
facenet-pytorch==2.5.3
mtcnn==0.1.1
numpy==1.24.4
Pillow==10.1.0
dlib==19.24.2